talk-with-docs-starter2-server/pyproject.toml
[project]
name = "talk-with-docs-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Brian H. Hough (@BrianHHough)"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "boto3 (>=1.38.35,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false